+++ packages/dynamics-lib/src/entities/contact.entity.js
@@ -32,7 +32,8 @@ export class Contact extends BaseEntity {
       },
       preferredMethodOfNewsletter: { field: 'defra_preferredmethodofnewsletter', type: 'optionset', ref: 'defra_preferredcontactmethod' },
       preferredMethodOfReminder: { field: 'defra_preferredmethodofreminder', type: 'optionset', ref: 'defra_preferredcontactmethod' },
-      postalFulfilment: { field: 'defra_postalfulfilment', type: 'boolean' }
+      postalFulfilment: { field: 'defra_postalfulfilment', type: 'boolean' },
+      obfuscatedDob: { field: 'defra_obfuscated_dob', type: 'string' }
     }
   }))
 
@@ -227,7 +228,7 @@ export class Contact extends BaseEntity {
 
   /**
    * Whether the user has requested a licence by post
-   * @type {GlobalOptionSetDefinition}
+   * @type {boolean}
    */
   get postalFulfilment () {
     return super._getState('postalFulfilment')
@@ -235,5 +236,17 @@ export class Contact extends BaseEntity {
 
   set postalFulfilment (postalFulfilment) {
     super._setState('postalFulfilment', postalFulfilment)
+  }
+
+  /**
+   * The obfuscated date of birth
+   * @type {string}
+   */
+  get obfuscatedDob () {
+    return super._getState('obfuscatedDob')
+  }
+
+  set obfuscatedDob (obfuscatedDob) {
+    super._setState('obfuscatedDob', obfuscatedDob)
   }
 }