+++ packages/gafl-webapp-service/src/pages/order-complete/order-complete/route.js
@@ -2,7 +2,7 @@ import pageRoute from '../../../routes/page-route.js'
 
 import Boom from '@hapi/boom'
 import { COMPLETION_STATUS, FEEDBACK_URI_DEFAULT } from '../../../constants.js'
-import { ORDER_COMPLETE, NEW_TRANSACTION, LICENCE_INFORMATION } from '../../../uri.js'
+import { ORDER_COMPLETE, NEW_TRANSACTION, LICENCE_DETAILS } from '../../../uri.js'
 import { displayStartTime } from '../../../processors/date-and-time-display.js'
 import * as mappings from '../../../processors/mapping-constants.js'
 import { nextPage } from '../../../routes/next-page.js'
@@ -40,7 +40,7 @@ const getData = async request => {
     uri: {
       new: NEW_TRANSACTION.uri,
       feedback: process.env.FEEDBACK_URI || FEEDBACK_URI_DEFAULT,
-      licenceInformation: LICENCE_INFORMATION.uri
+      licenceDetails: LICENCE_DETAILS.uri
     }
   }
 }