+++ packages/sales-api-service/src/__mocks__/test-data.js
@@ -24,6 +24,7 @@ const optionSetData = JSON.parse(readFileSync(optionSetDataPath, { encoding: 'UT
 
 const testExecutionTime = moment()
 export const MOCK_PERMISSION_NUMBER = '11100420-2WT1SFT-KPMW2C'
+export const MOCK_OBFUSCATED_DOB = '87200001013460'
 export const MOCK_START_DATE = testExecutionTime.toISOString()
 export const MOCK_END_DATE = testExecutionTime.add(1, 'year').toISOString()
 export const MOCK_ISSUE_DATE = testExecutionTime.toISOString()
@@ -43,7 +44,8 @@ export const mockContactPayload = () => ({
   preferredMethodOfConfirmation: 'Text',
   preferredMethodOfNewsletter: 'Email',
   preferredMethodOfReminder: 'Letter',
-  postalFulfilment: true
+  postalFulfilment: true,
+  obfuscatedDob: MOCK_OBFUSCATED_DOB
 })
 
 export const mockContactWithIdPayload = () => ({