+++ packages/sales-api-service/src/services/__tests__/permission-helper.spec.js
@@ -1,5 +1,5 @@
-import { logStartDateError } from '../permission-helper.js'
 import moment from 'moment'
+import { logStartDateError } from '../permission-helper.js'
 
 const consoleError = console.error
 
@@ -20,7 +20,10 @@ describe('logStartDateError', () => {
       issueDate: '2021-08-10T14:05:54Z'
     }
     logStartDateError(samplePermission)
-    expect(console.error).toHaveBeenCalled()
+    expect(console.error).toHaveBeenCalledWith(
+      'permission start date before issue date: ',
+      samplePermission
+    )
   })
 
   it('logs if start date is before current time, if no issue date is provided', () => {
@@ -28,10 +31,13 @@ describe('logStartDateError', () => {
       startDate: moment().subtract(5, 'hours').toISOString()
     }
     logStartDateError(samplePermission)
-    expect(console.error).toHaveBeenCalled()
+    expect(console.error).toHaveBeenCalledWith(
+      'permission start date before current time: ',
+      samplePermission
+    )
   })
 
-  it('doesn\'t log if start date is after issue date', () => {
+  it("doesn't log if start date is after issue date", () => {
     const samplePermission = {
       startDate: '2021-08-10T14:35:54Z',
       issueDate: '2021-08-10T14:05:54Z'
@@ -40,9 +46,18 @@ describe('logStartDateError', () => {
     expect(console.error).not.toHaveBeenCalled()
   })
 
-  it('doesn\'t log if start date is after current date', () => {
+  it("doesn't log if start date is after current date", () => {
     const samplePermission = {
       startDate: moment().add(30, 'minutes').toISOString()
+    }
+    logStartDateError(samplePermission)
+    expect(console.error).not.toHaveBeenCalled()
+  })
+
+  it("doesn't log if it's a POCL import", () => {
+    const samplePermission = {
+      startDate: moment().subtract(5, 'hours').toISOString(),
+      dataSource: { id: 910400000 }
     }
     logStartDateError(samplePermission)
     expect(console.error).not.toHaveBeenCalled()