+++ packages/sales-api-service/src/services/permission-helper.js
@@ -1,8 +1,11 @@
 import moment from 'moment'
+const POCL_DATA_SOURCE_ID = 910400000
 
 export const logStartDateError = permission => {
   const { startDate, issueDate } = permission
-  if (moment(startDate).isBefore(issueDate || undefined)) {
-    console.error('permission start date before current time: ', permission)
+  const startDateBeforeTargetDate = moment(startDate).isBefore(issueDate || undefined)
+  const isPOCLImport = permission?.dataSource?.id === POCL_DATA_SOURCE_ID
+  if (startDateBeforeTargetDate && !isPOCLImport) {
+    console.error(`permission start date before ${issueDate ? 'issue date' : 'current time'}: `, permission)
   }
 }